Refactor: move math engine to core/ and update imports. Added __init__.py in core/
