+++ core/plot_settings.py
