+++ version.py
@@ -1,3 +1,1 @@
-
-
 __version__ = '1.0.0'
